+++ src/components/HotLiItem/index.module.less
@@ -1,0 +1,84 @@
+.hot__li {
+  background: #fff;
+  margin: 20px 0 0;
+  border-radius: 25px;
+  padding: 20px 50px;
+  .hot__detail {
+    display: flex;
+    .hot__country {
+      flex: 1;
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      justify-content: center;
+      .hot__img {
+        height: 50px;
+        width: 50px;
+        border-radius: 50%;
+      }
+      .hot__name {
+        flex: 1;
+        font-size: 26px;
+        display: flex;
+        align-items: center;
+        padding: 8px 0 0 0;
+      }
+    }
+    .hot__score {
+      width: 100px;
+      display: flex;
+      font-size: 34px;
+      font-weight: bold;
+      justify-content: center;
+      align-items: center;
+    }
+  }
+  .hot__progress {
+    width: 100%;
+    margin: 5PX 0;
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    .hot__percent {
+      flex: 1;
+      font-size: 26px;
+      color: #999;
+    }
+    .hot__part {
+      height: 14px;
+      width: 100%;
+      background: #ddd;
+      display: flex;
+      border-radius: 20px;
+      margin: 3PX 0 5PX;
+      &__p1 {
+        background: #cccccc;
+        border-radius: 20px 0 0 20px;
+      }
+      &__p2 {
+        background: #dddddd;
+      }
+      &__p3 {
+        flex: 1;
+        background: #cb3131;
+        border-radius: 0 20px 20px 0;
+      }
+    }
+  }
+  .hot__note {
+    display: flex;
+    height: 50px;
+    font-size: 24px;
+    justify-content: flex-start;
+    align-items: center;
+    .hot__time {
+      flex: 1;
+      color: #999;
+    }
+    .hot__like {
+      height: 40px;
+      width: 40px;
+      float: right;
+    }
+  }
+}
